Modularized code in MangaReader.swift
