+++ Mango (Social Manga)/Extensions/MangaReader UICollectionViewDelegateFlowLayout.swift
@@ -1,0 +1,37 @@
+//
+//  MangaReader UICollectionViewDelegateFlowLayout.swift
+//  Mango (Social Manga)
+//
+//  Created by Blake Harrison on 10/15/18.
+//  Copyright © 2018 Blake Harrison. All rights reserved.
+//
+
+import UIKit
+
+//MARK: - Extensions
+extension MangaReader: UICollectionViewDelegateFlowLayout {
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+        
+        return 0
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        
+        let frameSize = collectionView.frame.size
+        return CGSize(width: frameSize.width - 0, height: frameSize.height)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        
+        return UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        if self.navigationController?.isNavigationBarHidden == false {
+            self.navigationController?.setNavigationBarHidden(true, animated: true)
+        } else if self.navigationController?.isNavigationBarHidden == true {
+            self.navigationController?.setNavigationBarHidden(false, animated: true)
+        }
+    }
+}
